fix(plugins): lazy-require audio-transcribe zod modules so discovery accepts it
audio-transcribe was rejected during plugin discovery — its entry graph
statically imported `zod` (`import { z } from 'zod'`), and the discovery
scanner forbids bare-module imports in the entry graph. Rejected plugins land
in discovery `errors`, never get registered, and so never appear in the
settings-UI approval list (only 3 of 4 plugins showed).

Mirror the task-provider-kaneo pattern: make index.ts a self-contained entry
(local structural types, no bare or out-of-plugin static imports) that loads
the real implementation via `import.meta.require('./runtime.js')`. The scanner
treats require-mode files as opaque, so the zod import inside runtime.ts is no
longer part of the scanned static graph.

- index.ts: slim require-based entry; activate() delegates to runtime
- runtime.ts (new): the zod schema + execute/transform logic + registration
- knip.jsonc: ignore the dynamically-required runtime.ts and the transcription
  helper it statically imports, matching the existing kaneo require-bridge entries

Discovery now reports 4 plugins / 0 errors; all 55 audio-transcribe tests pass.
